requirements.txt
click>=8.1.6
tzdata
tabulate
pyfzf_iter
